dovetail: add core support
This is the basic interface layer for hosting a high-priority resident
co-kernel, aka "dovetailing".

The co-kernel leverages the interrupt pipeline ability to turn common
device IRQs into pseudo-NMIs from the perspective of the in-band code
(CONFIG_IRQ_PIPELINE), to achieve very low latency for the co-kernel
running on the oob stage.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>

dovetail: Fix detection of prctl based compat syscalls

In case a compat application is requesting a dovetail service the
syscall number will not be equal to __NR_prctl (= the x86_64 syscall
number).

The same problem exists on the arm64 architecture where the problem was
already addressed. Apply the same to x86.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>

dovetail: arch: Fix fetching of the first syscall argument from compat tasks

On x86 there is a small difference in the ABI (= register) where the
first syscall argument is supplied between the x86_64 (native) and
compat (i386) case.

On x86_64 the DI register is being used, while i386 used the BX
register.

The same problem exists for further arguments - handled by
syscall_get_arguments() - as well. syscall_get_arg0() is dovetail
specific and is now in sync with syscall_get_arguments().

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>

dovetail: adjust core support for 6.14

Fixes the following splat while trying to suspend / resume from disk:

[   17.921496] ------------[ cut here ]------------
[   17.921502] WARNING: CPU: 3 PID: 33 at arch/x86/mm/tlb.c:516 switch_mm_irqs_off+0x39d/0x4f0
[   17.921512] Modules linked in:
[   17.921517] CPU: 3 UID: 0 PID: 33 Comm: cpuhp/3 Not tainted 6.14.0-rc7+ #82
[   17.921521] Hardware name: QEMU Standard PC (Q35 + ICH9, 2009), BIOS 1.16.2-debian-1.16.2-1 04/01/2014
[   17.921523] IRQ stage: Linux
[   17.921526] RIP: 0010:switch_mm_irqs_off+0x39d/0x4f0
[   17.921531] Code: 06 75 de 65 c6 05 e6 82 db 7e 00 e9 77 fd ff ff 48 8b 05 3e 79 18 01 b9 49 00 00 00 48 89 c2 48 c1 ea 20 0f 30 e9 fb fc ff ff <0f> 0b e9 9f fc ff ff 0f 0b e9 4f fd ff ff 65 48 c7 05e
[   17.921533] RSP: 0018:ffffc90000133de8 EFLAGS: 00010202
[   17.921537] RAX: 0000000000000001 RBX: ffffffff8270e000 RCX: 0000000000000207
[   17.921539] RDX: 0000000000000206 RSI: ffffffff8270e000 RDI: ffff8880065ae3c0
[   17.921541] RBP: ffff88800378ab80 R08: ffff88803e79c5c8 R09: 000000000000019d
[   17.921543] R10: ffff88803e79c5a0 R11: 0000000000000001 R12: ffff8880065ae3c0
[   17.921545] R13: 0000000000000000 R14: 0000000000000003 R15: ffff88803e79c5c8
[   17.921558] FS:  0000000000000000(0000) GS:ffff88803e780000(0000) knlGS:0000000000000000
[   17.921574] CS:  0010 DS: 0000 ES: 0000 CR0: 0000000080050033
[   17.921577] CR2: 000055ed7e138e28 CR3: 0000000003ea8000 CR4: 00000000003506f0
[   17.921579] Call Trace:
[   17.921622]  <TASK>
[   17.921625]  ? __warn+0x85/0x140
[   17.921630]  ? switch_mm_irqs_off+0x39d/0x4f0
[   17.921634]  ? report_bug+0x164/0x190
[   17.921641]  ? handle_bug+0x65/0xd0
[   17.921645]  ? exc_invalid_op+0xb2/0xd0
[   17.921648]  ? asm_exc_invalid_op+0x1a/0x20
[   17.921653]  ? switch_mm_irqs_off+0x39d/0x4f0
[   17.921657]  ? __pfx_sched_cpu_wait_empty+0x10/0x10
[   17.921661]  sched_cpu_wait_empty+0x93/0xc0
[   17.921665]  cpuhp_invoke_callback+0x114/0x4b0
[   17.921669]  cpuhp_thread_fun+0xee/0x190
[   17.921674]  ? __pfx_smpboot_thread_fn+0x10/0x10
[   17.921678]  smpboot_thread_fn+0xdd/0x1d0
[   17.921683]  kthread+0xf1/0x1e0
[   17.921689]  ? __pfx_kthread+0x10/0x10
[   17.921693]  ret_from_fork+0x34/0x50
[   17.921698]  ? __pfx_kthread+0x10/0x10
[   17.921701]  ret_from_fork_asm+0x1a/0x30
[   17.921717]  </TASK>
[   17.921719] ---[ end trace 0000000000000000 ]---

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>