evl/thread: open code transition to in-band context
Move the code handling the transition from OOB to in-band from
evl_thread_suspend() to evl_switch_inband().

Rationale: this code is 100% specific to evl_switch_inband() which is
the only allowed caller. This move simplifies the implementation of
evl_suspend_thread() while centralizing all the in-band transition
logic in a single place.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>