evl/clock: do not fast-forward on tick overrun
If a tick is overrunning, assume the best to do is to signal the next
event immediately instead of advancing to the next release point past
the current date, hoping that we are not outpaced by too short a
period. In the latter case, we are toast and would lock up regardless
of the implementation anyway.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>