evl/thread: ensure all blocking bits are lifted on kick
We must unblock _and_ release a thread which is being kicked out the
out-of-band stage (*), since these conditions are cumulative. As a
consequence, get rid of the pointless return value of
evl_wakeup_thread() and evl_unblock_thread().

(*) the additional round-robin effect we may get from
evl_release_thread() for the kicked thread is a non-issue, since it is
eventually going to switch inband anyway, so this is still correct
priority-wise.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>