evl/mutex: refcount mutex owners the kernel-side knows about
Since we keep a live pointer to the current mutex owner we may know
about for every locked mutex, we have to refcount the referenced
threads to prevent use-after-free bugs whenever a thread exits while
holding one or more mutexes.

This also fixes a refcounting issue with elements preventing proper
deletion of the device, as observed with the monitor-steal test.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>