evl/proxy: add read side
A proxy should be usable for reading from an in-band file without
encurring a stage switch, just like it is for the output direction.

These changes introduce the EVL_CLONE_INPUT, EVL_CLONE_OUTPUT creation
flags for a proxy, which can be used in a mutually exclusive manner,
or combined in order to have a bidirectional channel to the target
file.

As a by-product, the output direction gains in-band poll() support.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>