evl/sched: do not consider oob stall on enabling preemption
Checking the oob stall bit in __evl_enable_preempt() to block the
rescheduling is obsolete. It relates to a nested locking construct
which is long gone, when the evl_spinlock managed the preemption count
and the big lock was still in, i.e.:

lock_irqsave(&ugly_big_lock, flags)  /* stall bit raised */
	evl_spin_lock(&inner_lock);  /* +1 preempt */
	   wake_up_high_prio_thread();
	evl_spin_unlock(&inner_lock); /* -1 preempt == 0, NO schedule because stalled */
unlock_irqrestore(&ugly_big_lock, flags) /* stall bit restored */

This was a way to prevent a rescheduling to take place inadvertently
while holding the big lock.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>