evl/file: convert file table lock to hard lock
The data protected by the file table lock is frequently accessed from
the in-band stage where holding it with hard irqs off is
required. Therefore we would not benefit in the common case from the
preemption disabling feature we are going to add to the EVL-specific
spinlock. Make it a hard lock to clarify the intent.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>