evl/lock: add preemption tracking
An EVL lock is now distinct from a hard lock in that it tracks and
disables preemption in the core when held.

Such spinlock may be useful when only EVL threads running out-of-band
can contend for the lock, to the exclusion of out-of-band IRQ
handlers. In this case, disabling preemption before attempting to grab
the lock may be substituted to disabling hard irqs.

There are gotchas when using such type of lock from the in-band
context, see comments in evl/lock.h.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>