evl/mutex: simplify the release path
There is no point in duplicating the logic which transfer_ownership()
already provides.  We may unconditionally rely on this routine for
either passing the lock to the next waiter in line, or marking the
lock as free, updating the fast handle accordingly in both cases.

NOTE:

- FLCLAIM and FLCEIL are manipulated from kernel space only.

- FLCEIL is only meaningful to a locked mutex, we may and have to
clear it unconditionally on release. On transfer, the status of this
bit is decided based on the mutex type (i.e. PCP or PIP-enabled).

Signed-off-by: Philippe Gerum <rpm@xenomai.org>