evl/mutex: disentangle locking dependencies
This is a preparation patch for introducing a new locking model which
will address a truckload of ABBA issues already detected by
lockdep. This is not a fix per se, but several changes aimed at
disentangling some existing locking constructs when it is
straightforward, i.e. the low hanging fruits first. For this reason,
this is one bulk patch waiting for the whole locking approach to be
revised in the following patch series.

Namely,

- We don't have to - and must not - hold the last and next owners'
locks simultaneously when transferring ownership of a mutex. The state
of the thread dropping the mutex should be updated independently
first, before ownership is passed to the next waiter in line.
Removing this useless lock nesting fixes an ABBA issue in the release
path.

- In-band ownership detection does not require to hold curr->lock. Do
the detection before grabbing this lock.

- Untangle the locking constructs when about to sleep on the mutex, so
that we can group the operations for the current and owner threads,
both for acquisition and release.

- Introduce a safe, double ordered locking helper for threads, to
escape the ABBA issue when we need to lock the current and mutex owner
thread concurrently.  Threads are locked by increasing address order,
which garantees a fixed locking sequence between them.

- We do not have to hold a lock on the originator of the locking
request when traversing the dependency chain for detecting a
deadlock. The latter only requires to match such thread by address in
any depending wait channel.  Drop this constraint from the
follow_depend() and reorder_wait() handlers.

- When walking through all waiters for a mutex in
evl_follow_mutex_depend(), we don't have to hold their respective
thread->lock since they would have to leave the wait list they belong
before they can go stale, which they cannot in the first place since
we hold mutex->lock while iterating. Eventually, drop the useless
locking of waiter->lock.

- We don't need to hold curr->lock across the call validating the lock
dependency chain. Split the thread lock release code in order to drop
curr->lock as early as possible, before check_lock_chain() is called.

- In transfer_ownership(), we must guard against the new owner going to
pend on a resource while removing it from the wait list.

- Flatten the deadlock detection logic when traversing the lock chain.

- There is no point in carrying over the (mutex, owner) owner tuple as
argument in places where mutex->owner == owner by
construction. Simplify some call signatures accordingly.