evl/wait: share common wait-schedule code
Since we have all the needed information for maintaining a wait list
into the wait channel descriptor embedded into the waitqueue, we may
share the wait-schedule code with other parts of the core,
particularly the mutex support.

Export this common code as __evl_wait_schedule() accordingly, updating
the wait-related tracepoints to operate on wait channels directly,
instead of waitqueues.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>