evl/mutex: remove invalid assertion
We may enter adjust_owner_boost() with no boosters because we did not
hold owner->lock on entry, therefore this owner might have dropped all
of them in the meantime if not current on the local CPU.

Drop the assertion marking this condition as an issue. However, check
that T_BOOST is cleared if the latter is met.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>