evl/sched: tp: deny changing plan on busy runqueues
Unconditionally switching the remaining TP threads on a runqueue to
the SCHED_FIFO class before swapping the TP schedule plan is an
arbitrary decision which may have unwanted side-effects, even papering
over a consistency issue in the application.

Besides, the implementation had the locking wrong in a couple of
places, such as failing to grab the thread lock before calling
evl_set_thread_schedparam_locked().

Change the logic by detecting the case and simply denying the change
instead, returning -EBUSY to the caller. User must ensure to update
the schedule plan consistently, i.e. when no thread is still running
the previous one.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>