evl/sched: introduce sched_name() handler
Some policies may advertise their name differently depending on the
thread for which it is queried, in which case reporting the policy's
global name would be inaccurate. E.g. SCHED_FIFO + EVL_T_RRB should be
reported as round-robin, not fifo, although the former actually sits
on the latter.

Introduce the per-policy sched_name() handler, which can deal with
such details based on a thread context.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>