net: dovetail: fix use-after-free in __sock_release()
Although related, socket and sock struct instances have distinct
lifetimes. As a consequence, the former might have been freed already
when the latter is released eventually. The wrong assumption that both
would have the same lifetime caused __sock_release() to spuriously
refer to a deallocated sock struct in a call to the sock_oob_detach()
handler (caught by Kasan).

To fix this, call sock_oob_detach() from __sk_destruct() instead,
which matches the semantics of the former for the oob side.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>