evl/net: fix endianness in protocol matching
The ->attach() handler of a protocol descriptor receives the protocol
number in the wrong (native) endianness. This illustrates a general
issue with the current code: we should not be dealing with protocol
identifiers in network byte order internally, this should be
restricted to external interface code.

Switch all protocol identifiers to native endianness except when
dealing with external interfaces.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>