genirq: irq_pipeline: enable local_irq_*_full() on both stages
As their names suggest, these calls disable/restore both the hardware
mask and stall bit when running in-band, so as to create a full
interrupt-free section while keeping the stall bit in sync. Extend
these semantics to the oob stage, so that only the hardware state is
affected, since we have no stall bit there.

We already have the test_and_lock_stage() and unlock_stage() calls
available for this purpose, remap local_irq_*_full() to those ones.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>