evl/monitor: enable in-band read/write ops on event mask
Event flags provide simple and flexible semantics for notifying
threads on various conditions. Let's extend this to in-band threads
including non-EVL ones, so that an EVL flag group may be used as a
cross-stage synchronization feature as well.

To this end, we now provide support for read(2), write(2) and poll(2)
syscalls on EVL event flags. In contrast to the oob forms for
notifying and waiting on a flag group via oob_ioctl(), the in-band
forms only provide support for unicast, wait-any operations.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>