evl: introduce random number generators
Random number generation is a requirement for some real-time
applications. Unfortunately, the WCET and jitter of the generators
available from the 'random' chardev implementation are not acceptable
for the typical EVL use cases.

To address this, we introduce three generators producing u8, u16 and
u32 random values. These generators can be read both from the
in-band and out-of-band I/O interfaces.

Signed-off-by: Lionel Minne <lionel.minne@exail.com>
Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>