net: dovetail: rename sock_oob_detach() to sock_oob_destroy()
This call is about dismantling a network socket, freeing all its
resources. This is the second part of a two-step deletion process
which may happen asynchronously due to refcounting, paired with the
release of the encapsulating kernel socket.

Rename sock_oob_detach() to sock_oob_destroy() since this is not a
strict converse of sock_oob_attach().

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>