net: dovetail: use file binding to detect oob socket
sock->sk might be stale while the encpasulating BSD socket is not, so
dereferencing sock->sk is not always safe.

Since the companion kernel must turn on oob capability for the file
struct associated to an oob-enabled BSD socket, we can use this
information to figure out whether the network representation of the
socket is oob capable as well.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>