evl/net: socket: handle release event
sock_oob_release() is called when the file struct associated to the
socket is closed, which unlike sock_oob_destroy() may block.

Move the calls to evl_release_file() and evl_pass_crossing() to the
release hook, which fixes an issue with sock_oob_destroy() attempting
to block while passing the crossing in a RCU callback context.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>