evl/net: dev: enable oob for any netdevice
So far, turning on oob diversion was available to VLAN devices
only. The rationale behind this restriction was that the VLAN id could
be used as a discriminator for oob traffic, feeding the EVL netstack
instead of the inband one.

In preparation for supporting per-device eBPF module in the netstack
which would serve as a user-defined traffic filter, we can lift most
of this general restriction.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>