dovetail: Account for invalid syscall number
At least on x86, the kernel may have decided to invalidate the syscall
number while processing it. Avoid that we process this -1 as legacy
syscall.

This was a long-pending issue in the whole dovetail series, now revealed
by running auditd on one of the test systems. Syscall auditing was
eventually catching an unbalanced tracing, caused by a premature syscall
exit due to this bug.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>