dovetail: net: rename oob_data to sk_oob_ctx in struct sock
The notion of "oob data" is used across several kernel subsystems,
particularly in the networking area. Dovetail's notion of out-of-band
relates to the execution context of a CPU on the high priority stage
of the interrupt pipeline instead (i.e. the one which runs
asynchronously to the regular scheduling of operations).

Mitigate the confusion in struct file by moving from the way too
generic "data" to "context", abiding by the common naming scheme there
in the same move, i.e. s/oob_data/sk_oob_ctx/.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>