genirq: irq_pipeline: fix hardirq context detection (irqctx)
Each incoming IRQ passed to handle_irq_desc() is checked for calling
context correctness, in order to detect invalid calls from
non-interrupt context for interrupts marked as
IRQD_HANDLE_ENFORCE_IRQCTX.

When pipelining is enabled, the IRQ context spans from the early
handling from the pipeline entry (i.e. in_pipeline()) up to the
completion point of the full interrupt path, which includes the
oob->in-band deferral path.

The test checking for the latter was mistakenly inverted, causing the
interrupt to be dismissed if bearing the IRQD_HANDLE_ENFORCE_IRQCTX
flag.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>