evl/net: provide neighbour solicitation service
We need a way to guarantee zero-delay output to a given IP
destination, which implies to resolve all routing and neighbour
solicitation chores beforehand, so that we won't have to pass on the
initial output to the inband stack, waiting for the oob front caches
(IP routes and ARP) to be updated on the fly.

To this end, ioctl(EVL_SOCKIOC_SOLICIT) is introduced for any IPv4 EVL
socket (UDP so far). This request waits for the EVL front caches to
receive the information needed to complete an output strictly from the
oob stage, triggering a route resolution and neighbour solicitation as
required.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>