net: dovetail: finalize alls skbs to head cache
Abusing the semantics of SKB_FCLONE_UNAVAILABLE to determine whether a
skb should go to the NAPI cache makes no sense, this is fragile at
best, and likely wrong with real clones. Since the NAPI cache will be
refilled from the head cache as needed when depleted, we can release
all finalized skbs to the latter.

This change somewhat partially disables the NAPI fast cache for
devices operating as oob ports when releasing inband-originated skbs
after transmission (i.e. when cleaning the TX ring). This said, that
is acceptable since mixing inband and oob traffic through the same
device is not recommended, although definitely possible, and we
certainly/obviously don't expect high traffic load on the inband side
in such a case.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>