net: dovetail: restrict napi_complete_oob() to inband polling completion
napi_complete_oob() is called by napi_complete_done() to start
processing the ingress packets received from a device which is not
oob-capable. Make sure to restrict this call to the latter case,
i.e. when napi_complete_done() is running inband exclusively.

The out-of-band net stack is expected to schedule its RX thread
immediately on the oob receive path.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>