net: dovetail: fold napi_complete_oob() into napi_schedule_oob()
So far, napi_schedule_oob() was invoked from the oob stage
exclusively, aimed at queuing a NAPI struct for RX handling by the
companion core from the oob stage.

On the other hand, napi_complete_oob() was expected to schedule the RX
handling of the packets accepted (in-band) by netif_deliver_oob() from
a non oob-capable device diverting traffic to the companion core, so
that it passes those packets to the proper protocol handlers in the
oob netstack.

We can ask the companion core to fold the logic of napi_complete_oob()
into napi_schedule_oob() by checking the current execution stage,
requiring a single hook for both cases.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>