evl/net: skb: save cached copy of netdev address
We need to keep a pointer to the originating device for the release
path (__free_evl_skb), since a regular protocol layer (e.g. UDP) might
use skb->dev_scratch which would cause skb->dev to be lost.

So we only rely on skb->dev as long as it is safe to do so, switching
to EVL_NET_CB(skb)->dev afterwards.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>