evl/net: skb: move owning device to shared info
We cannot safely use the control block to store the owning device
pointer specifically when the NIC driver is not oob-capable, in which
case we'd have to hand the buffer over to the inband stack
(dev_queue_xmit()) where the cb might be used for other purpose.

Move the device pointer from the cb to the shared oob information
descriptor available in every skb.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>