evl/net: enable oob traffic via loopback devices
This patch allows for enabling 'lo' as an oob networking port. When
enabled, the loopback device accepts back any oob traffic which it
receives.

As a consequence, packets sent to 127.0.0.1 from the oob stage will be
delivered to the 'lo' oob port. Conversely, in-band traffic sent to
the loopback interface will be delivered to the in-band
stack. Loopback devices won't generally support VLANs, however we
still allow incoming in-band traffic to be directed to the oob port by
mean of a BPF program.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>