igb: fix operations in primary interrupt context (in-band vs oob)
The set of operations required for proper IRQ handling depends on the
current stage (in-band vs oob). Because IRQ forwarding is used when
oob mode is enabled for a device, most handlers may run twice for a
single interrupt, first from the oob stage, next from the in-band
stage as a result of IRQ forwarding by the former.

As a result, we have to differentiate between in-band execution due to
forwarding, or because oob diversion is turned off for the
device. Meanwhile, make sure to reset the [TS]ICR cache appropriately
when an interrupt request ends.

Tot this end, introduce igb_in_primary_irqctx() and igb_eoi().

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>