evl/net: ipv4: use ARP pseudo-entry for lo device
Using a regular entry in the ARP cache for the special loopback device
is an issue because it is affected when flushing the cache, although
it should not. This creates weird corner cases such as the requirement
to solicit the loopback address in order to restore its entry after a
cache flush.

Since we are going to solve the same kind of issue with the upcoming
broadcast address, let's provide ARP pseudo-entries, which are legit,
refcounted but uncached ARP entries, accessed via a global pointer
from the net namespace.

Fix up evl_net_get_arp_entry() to specifically recognize the loopback
address, returning its pseudo entry when matched instead of looking up
into the ARP cache.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>