evl/net: add support for network interface taps
Feed the in-band network interface taps with ingress traffic we
accept, and egress traffic we produce. The scheme works with or
without oob support from the NIC device the netstack deals with.

This addition enables common traffic sniffers to capture all traffic
going through EVL netstack (wireshark, tcpdump and friends).

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>