net: dovetail: add tap feeder on the receive side
We need an RX counterpart to dev_queue_xmit_nit() for use by the oob
netstack, in order to feed the in-band sniffers listening to ETH_ALL
packets.

To this end, dev_queue_recv_nit() allows a companion core to feed the
in-band taps with (a clone of) every skb it handles, so that common
packet capture tools can see the ingress oob traffic as well.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>