evl/net: ethernet: fix buffer leak on transmit error
We need evl_net_ether_transmit() to drop all skbs starting from the
heading one on error, since we have the information about the precise
failing point right there in order to release any orphaned fragments,
which the caller does not.

Fix the single caller in the UDP implementation accordingly.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>