evl/net: arp: resolve broadcast address via a pseudo-entry
Since the (limited) broadcast address is an IPv4 constant, we can
build a pseudo-entry for it, just like for the loopback one.

So, if evl_net_solicit() finds a device with an active oob port which
supports broadcast, then passing this device along with the broadcast
address to evl_net_get_arp_entry_or_pseudo() will build a valid
ARP pseudo-entry for oob broadcasting, e.g. of UDP frames.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>