evl/net: socket: drop get_netif handler
We have no more users of this handler, which never made much sense
anyway, since we use the binding information locally in the protocol
layers where we always know where to find it. IOW, there should not be
cross-layer requests for binding information which would require such
abstraction/indirection.

Drop this cruft.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>