evl/net: socket: deny solicitation if bound device is invalid
So far, we passed a null device pointer to the solicitation service in
case the device bound to the socket did exist but was not an oob
port. This could yield an ambiguous error status, such as receiving
-ENETUNREACH from ip_output_route(), although the existing device
should be usable for in-band routing to the given address.

Fix this by denying the solicitation entirely if the bound device is
not an oob port, returning -ENODEV, which is the conventional error
code in this situation.

In addition, ensure that evl_net_ipv4_solicit() also declines to
search for a route if given a valid device which is not an oob port,
returning -ENODEV as well.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>