clocksource/hyper-v: irq_pipeline: Enable pipelined clock events
Make the clock event around the STIMER0 pipeline safe. The
implementation is following the same pattern that we already have in
place for the x86 LAPIC interrupt / clock event.

To get the proxy tick working we need to link the hyperv clock event
device with the STIMER0 vector. This part is also following the pattern
of the x86 lapic interrupt. Both have a fixed arch specific vector
assigned.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>