evl/net: ipv4: fix probing of connected/stale neighbour on solicit
Upon solicitation, we might find a connected neighbour into the inband
ARP table which is not present in our front cache, or a stale
neighbour entry, which would yield the following issues:

- for an entry in connected state but not present in our front cache
(*), the inband ARP logic would not send a probe request due to the
state, which would in turn prevent the peer to be added to our front
cache since it listens for successful probes from the inband ARP
management.

- for an entry in stale state - therefore absent from our front cache
by definition - the inband ARP logic would not send a probe request
either.

Either way, this would cause evl_net_ipv4_solicit() to bail out with a
timeout error.

To fix this, force update our front cache with any connected neighbour
received from the routing process, instead of asking the inband ARP
management to probe for the peer. In the same move, make sure to
invalidate/expire stale entries to force probing instead.

(*) typically, this may happen if the front cache is forcibly flushed
but the inband ARP table is not.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>