e1000e: assume no oob traffic on oob mode switch
We may assume that no RX/TX oob traffic can happen while the {disable,
enable}_oob handlers are running. This MUST be enforced by the oob
netstack.

For this reason, synchronizing with the NAPI and/or oob TX is
pointless (the former may even cause the operation to hang
indefinitely).

Signed-off-by: Philippe Gerum <rpm@xenomai.org>