evl/net: skb: fix leak on clone
Cloning lost the oob marker set on the clone buffer obtained from the
out-of-band pool as a result of morphing the source into the
clone. This has resulted in clone buffers not being pushed back to the
originating pool on release, which ended up in depleting the pool
entirely.

Reset that marker appropriately after morphing.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>