evl/tick: fix in-band tick deferral
We cannot check for RQ_TDEFER, RQ_TSTOPPED or any other CPU-local flag
from a remote CPU for that matter, this is fundamentally unsound. By
doing so, the implementation is at risk of skipping a timer shot
unexpectedly when considering remote timers. Meanwhile, testing
RQ_TDEFER as a fallback check in order to figure out whether the
heading timer might be next to the deferred in-band one in the timer
queue is uselessly convoluted, we'd better know up front which is the
timer we programmed the next clock event for instead.

Let's refactor the in-band tick deferral logic in order to track the
timer for which we have a clock event pending at any point in time,
using this information to know whether some timer is the heading one
hardware-wise.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>